+++ bin/check.ml
@@ -23,9 +23,7 @@ let path2module p =
 
 let type_check load_path name sigs =
   let md = init_muc name in
-  let penv =
-    path2module name |> Utils.Sstr.singleton |> Typing.penv load_path
-  in
+  let penv = Utils.Sstr.singleton name |> Typing.penv load_path in
   let md = List.fold_left (Typing.type_sig_item penv) md sigs in
   md