+++ src/typing.mli
@@ -9,7 +9,6 @@
 (*                                                                  *)
 (********************************************************************)
 
-open Tast
 open Tmodule
 
 type parse_env
@@ -20,10 +19,6 @@ val penv : string list -> Utils.Sstr.t -> parse_env
     name `module_nm`. The paths in `load_paths` are to be used when searching
     for modules dependencies. *)
 
-val process_sig_item :
-  parse_env -> module_uc -> Uast.s_signature_item -> module_uc * signature_item
-(** `process_sig_item penv muc s` returns a new module under construction after
-    type checking `s` and the typed signature obtained from `s` and `muc`. *)
-
 val type_sig_item : parse_env -> module_uc -> Uast.s_signature_item -> module_uc
-(** the same as above but it drops the typed signature *)
+(** `type_sig_item penv muc s` returns a new module under construction after
+    type checking `s`. *)