+++ CHANGES.md
@@ -174,6 +174,8 @@
   [\#155](https://github.com/ocaml-gospel/gospel/pull/155)
 - Removed the `Pty_open` type constructor
   [\#109](https://github.com/ocaml-gospel/gospel/pull/109)
+- Added paths to gospel identifiers.
+  [\#396](https://github.com/ocaml-gospel/gospel/pull/396)
 
 
 # 0.1.0 (11-03-2021)